requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
Werkzeug
python-multipart
Jinja2